[SMA-277] Add module evoleq; Move common folder of api-data to evoleq
