+++ evoleq/detekt/detekt-baseline.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" ?>
+<SmellBaseline>
+    <ManuallySuppressedIssues>
+    </ManuallySuppressedIssues>
+    <CurrentIssues>
+    </CurrentIssues>
+</SmellBaseline>
