+++ evoleq/src/commonMain/kotlin/org/evoleq/ktorx/result/result-maths.kt
