+++ evoleq/src/commonMain/kotlin/org/evoleq/math/function.kt
