+++ evoleq/src/commonMain/kotlin/org/evoleq/math/is-double.kt
