+++ evoleq/src/commonMain/kotlin/org/evoleq/math/maths.kt
