+++ evoleq/src/commonMain/kotlin/org/evoleq/math/product.kt
