+++ evoleq/src/commonMain/kotlin/org/evoleq/math/state/state.kt
