+++ evoleq/src/commonMain/kotlin/org/evoleq/math/writer.kt
