+++ evoleq/src/commonMain/kotlin/org/evoleq/uuid/uuid.kt
