+++ evoleq/src/commonTest/kotlin/org/evoleq/math/is-double-test.kt
